Closes #1
Because this machine is only really useful on the server, it doesn't
make sense to have the callback option to return JSONP responses.

Additionally if you wish to use any of the options you are required
to use all the options (machinepack specification seems to only allow
one level of inputs for defining what is required, cannot set sub-fields
as not required but child sub-fields as required.).

So removing this entirely seems like the best course of action to
allow JSON only responses while being able to use the other more
useful options.